9-8-2020-08:46am

+++ app/Http/Controllers/Api/AddressController.php
@@ -22,38 +22,46 @@ class AddressController extends Controller
         ], 200);
     }
     
-    public function getAddressesDetails(){
-        return response()->json([
-            'countries' => CountryResource::collection(Country::all()),
-            'cities' => CityResource::collection(City::all()),
-        ], 200);
-    }
+    // public function getAddressesDetails(){
+    //     return response()->json([
+    //         'countries' => CountryResource::collection(Country::all()),
+    //         'cities' => CityResource::collection(City::all()),
+    //     ], 200);
+    // }
 
     public function store(Request $request){
         if(app('request')->header('Authorization') != null && Auth::guard('api')->check()){
+            if(app('request')->header('Authorization') == 'Bearer '.auth()->user()->api_token){
+                // dd(app('request')->header('Authorization'));
+                $this->validate($request, [
+                    // 'country_id' => 'required',
+                    // 'city_id' => 'required',
+                    'lang' => 'required',
+                    'long' => 'required',
+                    'address' => 'required',
+                    // 'postal_code' => 'required',
+                ]);
 
-            // dd(app('request')->header('Authorization'));
-            $this->validate($request, [
-                'country_id' => 'required',
-                'city_id' => 'required',
-                'address' => 'required',
-                'postal_code' => 'required',
-            ]);
+                $address = Address::create($request->all());
+                $address->update(['user_id' => auth()->user()->id]);
+                auth()->user()->addresses()->save($address);
 
-            $address = Address::create($request->all());
-            $address->update(['user_id' => auth()->user()->id]);
-            auth()->user()->addresses()->save($address);
+                if($address == null){
+                    dd('1');
+                    return response()->json([
+                        'error' => trans('api.error')
+                    ], 400);
+                }
 
-            if($address == null){
-                dd('1');
                 return response()->json([
-                    'error' => trans('api.error')
+                    'success' => trans('api.success')
+                ], 200);
+            }
+            else{
+                return response()->json([
+                    'error' => trans('api.unauthorized')
                 ], 400);
             }
-
-            return response()->json([
-                'success' => trans('api.success')
-            ], 200);
         }
         else{
             return response()->json([
@@ -64,34 +72,40 @@ class AddressController extends Controller
 
     public function update(Request $request){
         if(app('request')->header('Authorization') != null && Auth::guard('api')->check()){
-
-            $this->validate($request, [
-                'id' => 'required',
-                'country_id' => 'required',
-                'city_id' => 'required',
-                'address' => 'required',
-                'postal_code' => 'required',
-            ]);
-
-            $address = auth()->user()->addresses()->find($request->id);
-
-            if($address == null){
+            if(app('request')->header('Authorization') == 'Bearer '.auth()->user()->api_token){
+                $this->validate($request, [
+                    'id' => 'required',
+                    // 'country_id' => 'required',
+                    // 'city_id' => 'required',
+                    'lang' => 'required',
+                    'long' => 'required',
+                    'address' => 'required',
+                    // 'postal_code' => 'required',
+                ]);
+    
+                $address = auth()->user()->addresses()->find($request->id);
+    
+                if($address == null){
+                    return response()->json([
+                        'error' => trans('api.error')
+                    ], 400);
+                }
+    
+                $address->update($request->all());
+    
+                if(!$address){
+                    return response()->json([
+                        'error' => trans('api.error')
+                    ], 400);
+                }
+    
                 return response()->json([
-                    'error' => trans('api.error')
-                ], 400);
+                    'success' => trans('api.success')
+                ], 200);
             }
-
-            $address->update($request->all());
-
-            if(!$address){
-                return response()->json([
-                    'error' => trans('api.error')
-                ], 400);
-            }
-
             return response()->json([
-                'success' => trans('api.success')
-            ], 200);
+                'error' => trans('api.unauthorized')
+            ], 400);
         }
         else{
             return response()->json([
@@ -102,22 +116,26 @@ class AddressController extends Controller
 
     public function destroy(Request $request){
         if(app('request')->header('Authorization') != null && Auth::guard('api')->check()){
+            if(app('request')->header('Authorization') == 'Bearer '.auth()->user()->api_token){
+                $this->validate($request, ['id' => 'required']);
 
-            $this->validate($request, ['id' => 'required']);
+                $address = auth()->user()->addresses()->find($request->id);
 
-            $address = auth()->user()->addresses()->find($request->id);
+                if($address == null){
+                    return response()->json([
+                        'error' => trans('api.error')
+                    ], 400);
+                }
 
-            if($address == null){
+                $address->delete();
+
                 return response()->json([
-                    'error' => trans('api.error')
-                ], 400);
+                    'success' => trans('api.success')
+                ], 200);
             }
-
-            $address->delete();
-
             return response()->json([
-                'success' => trans('api.success')
-            ], 200);
+                'error' => trans('api.unauthorized')
+            ], 400);
         }
         else{
             return response()->json([
